moved non-api functions from resource_api to ld_util; added stub for ld_util_test
