+++ src/tests/unit_tests/util/ld_util_test.py
@@ -1,0 +1,49 @@
+from mockito import when, unstub, mock, verify
+
+def test_prepare_beng_uri():
+    try:
+        pass
+    finally:
+        unstub()
+
+def get_lod_resource_from_rdf_store():
+    try:
+        pass
+    finally:
+        unstub()
+
+def json_header_from_rdf_graph():
+    try:
+        pass
+    finally:
+        unstub()
+
+def json_iri_iri_from_rdf_graph():
+    try:
+        pass
+    finally:
+        unstub()
+
+def json_iri_lit_from_rdf_graph():
+    try:
+        pass
+    finally:
+        unstub()
+
+def json_iri_bnode_from_rdf_graph():
+    try:
+        pass
+    finally:
+        unstub()
+
+def is_public_resource():
+    try:
+        pass
+    finally:
+        unstub()
+
+def prepare_lod_resource_uri():
+    try:
+        pass
+    finally:
+        unstub()
